🔄 synced local 'skyvern/' with remote 'skyvern/'
<!-- ELLIPSIS_HIDDEN -->

> [!IMPORTANT]
> Introduces `UserDefinedError` for enhanced error handling in task failure summaries and updates related functions and templates.
>
>   - **Error Handling**:
>     - Introduces `UserDefinedError` in `errors.py` for custom error handling with fields `error_code`, `reasoning`, and `confidence_float`.
>     - Updates `MaxStepsReasonResponse` in `prompt_engine.py` to include a list of `UserDefinedError`.
>     - Modifies `summary_failure_reason_for_max_steps()` in `agent.py` to return `MaxStepsReasonResponse` with user-defined errors.
>   - **Prompt Template**:
>     - Updates `summarize-max-steps-reason.j2` to include user-defined errors in the JSON response.
>   - **Code Refactoring**:
>     - Removes `UserDefinedError` from `actions.py` and imports it from `errors.py` instead.
>     - Updates `AgentStepOutput` in `steps.py` to include a list of `UserDefinedError`.
>
> <sup>This description was created by </sup>[<img alt="Ellipsis" src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=Skyvern-AI%2Fskyvern-cloud&utm_source=github&utm_medium=referral)<sup> for 7535a62bcbf9732e1d149ab9139c4c44cb5f26f8. You can [customize](https://app.ellipsis.dev/Skyvern-AI/settings/summaries) this summary. It will automatically update as commits are pushed.</sup>

<!-- ELLIPSIS_HIDDEN -->
