+++ skyvern/utils/prompt_engine.py
@@ -4,6 +4,7 @@ import structlog
 from pydantic import BaseModel
 
 from skyvern.constants import DEFAULT_MAX_TOKENS
+from skyvern.errors.errors import UserDefinedError
 from skyvern.forge.sdk.prompting import PromptEngine
 from skyvern.utils.token_counter import count_tokens
 from skyvern.webeye.scraper.scraper import ElementTreeBuilder
@@ -28,6 +29,12 @@ class CheckDateFormatResponse(BaseModel):
 
 
 HTMLTreeStr = str
+
+
+class MaxStepsReasonResponse(BaseModel):
+    page_info: str
+    reasoning: str
+    errors: list[UserDefinedError]
 
 
 def load_prompt_with_elements(
