+++ skyvern/webeye/actions/actions.py
@@ -6,6 +6,7 @@ import structlog
 from litellm import ConfigDict
 from pydantic import BaseModel, Field
 
+from skyvern.errors.errors import UserDefinedError
 from skyvern.webeye.actions.action_types import ActionType
 
 LOG = structlog.get_logger()
@@ -17,15 +18,6 @@ class ActionStatus(StrEnum):
     skipped = "skipped"
     failed = "failed"
     completed = "completed"
-
-
-class UserDefinedError(BaseModel):
-    error_code: str
-    reasoning: str
-    confidence_float: float = Field(..., ge=0, le=1)
-
-    def __repr__(self) -> str:
-        return f"{self.reasoning}(error_code={self.error_code}, confidence_float={self.confidence_float})"
 
 
 class SelectOption(BaseModel):
