+++ skyvern/webeye/actions/models.py
@@ -6,8 +6,9 @@ from openai.types.responses.response import Response as OpenAIResponse
 from pydantic import BaseModel, ConfigDict
 
 from skyvern.config import settings
+from skyvern.errors.errors import UserDefinedError
 from skyvern.schemas.steps import AgentStepOutput
-from skyvern.webeye.actions.actions import Action, DecisiveAction, UserDefinedError
+from skyvern.webeye.actions.actions import Action, DecisiveAction
 from skyvern.webeye.actions.responses import ActionResult
 from skyvern.webeye.scraper.scraper import ScrapedPage
